+++ python-wrapper/tests/test_neo4j.py
@@ -12,8 +12,11 @@ from neo4j_viz.node import Node
 @pytest.fixture(scope="class", autouse=True)
 def graph_setup(neo4j_session: Session) -> Generator[None, None, None]:
     neo4j_session.run(
-        "CREATE (a:_CI_A {name:'Alice', height:20, id:42, _id: 1337, caption: 'hello'})-[:KNOWS {year: 2025, id: 41, source: 1, target: 2}]->"
-        "(b:_CI_A:_CI_B {name:'Bob', height:10, id: 84, size: 11, labels: [1,2]}), (b)-[:RELATED {year: 2015, _type: 'A', caption:'hej'}]->(a)"
+        "CREATE "
+        "  (a:_CI_A {name:'Alice', height:20, id:42, _id: 1337, caption: 'hello'})"
+        " ,(b:_CI_A:_CI_B {name:'Bob', height:10, id: 84, size: 11, labels: [1,2]})"
+        " ,(a)-[:KNOWS {year: 2025, id: 41, source: 1, target: 2}]->(b)"
+        " ,(b)-[:RELATED {year: 2015, _type: 'A', caption:'hej'}]->(a)"
     )
     yield
     neo4j_session.run("MATCH (n:_CI_A|_CI_B) DETACH DELETE n")
@@ -44,8 +47,9 @@ def test_from_neo4j_graph_basic(neo4j_session: Session) -> None:
         Node(
             id=node_ids[1],
             caption="_CI_A:_CI_B",
-            size=11,
+            size=None,
             properties=dict(
+                size=11,
                 labels=["_CI_A", "_CI_B"],
                 name="Bob",
                 height=10,
@@ -64,6 +68,41 @@ def test_from_neo4j_graph_basic(neo4j_session: Session) -> None:
         (node_ids[0], node_ids[1], "KNOWS"),
         (node_ids[1], node_ids[0], "RELATED"),
     ]
+
+
+@pytest.mark.requires_neo4j_and_gds
+def test_from_neo4j_graph_size_property(neo4j_session: Session) -> None:
+    # set a non parsable size property, by default it should not be picked up
+    neo4j_session.run("MATCH (n) SET n.size = 'banana'")
+
+    graph = neo4j_session.run("MATCH (a:_CI_A|_CI_B)-[r]->(b) RETURN a, b, r ORDER BY a").graph()
+
+    VG = from_neo4j(graph, size_property="height", node_radius_min_max=None)
+
+    assert {n.properties["name"]: n.size for n in VG.nodes} == {"Alice": 20, "Bob": 10}
+
+    VG = from_neo4j(graph, size_property=None, node_radius_min_max=None)
+
+    assert {n.properties["name"]: n.size for n in VG.nodes} == {"Alice": None, "Bob": None}
+
+
+@pytest.mark.requires_neo4j_and_gds
+def test_from_neo4j_graph_default_caption(neo4j_session: Session) -> None:
+    neo4j_session.run("MATCH (n) SET n.caption = 'my_caption' SET n.other_caption = 'other_caption'")
+
+    graph = neo4j_session.run("MATCH (a:_CI_A|_CI_B)-[r]->(b) RETURN a, b, r ORDER BY a").graph()
+
+    VG = from_neo4j(graph, node_caption=None, node_radius_min_max=None)
+
+    assert [n.caption for n in VG.nodes] == [None, None]
+
+    VG = from_neo4j(graph, node_caption="other_caption", node_radius_min_max=None)
+
+    assert [n.caption for n in VG.nodes] == ["other_caption", "other_caption"]
+
+    VG = from_neo4j(graph, relationship_caption="year")
+
+    assert {e.properties["type"]: e.caption for e in VG.relationships} == {"KNOWS": "2025", "RELATED": "2015"}
 
 
 @pytest.mark.requires_neo4j_and_gds
@@ -93,8 +132,8 @@ def test_from_neo4j_result(neo4j_session: Session) -> None:
         Node(
             id=node_ids[1],
             caption="_CI_A:_CI_B",
-            size=11,
             properties=dict(
+                size=11,
                 labels=["_CI_A", "_CI_B"],
                 name="Bob",
                 height=10,
@@ -230,9 +269,9 @@ def test_from_neo4j_graph_driver(neo4j_session: Session, neo4j_driver: Driver) -
         Node(
             id=node_ids[1],
             caption="_CI_A:_CI_B",
-            size=11,
             properties=dict(
                 labels=["_CI_A", "_CI_B"],
+                size=11,
                 name="Bob",
                 height=10,
                 id=84,